#8 getting an error

+++ Cefalo.farhadcodes-a-CP-blog.Service/Services/UserService.cs
@@ -9,6 +9,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Cefalo.farhadcodes_a_CP_blog.Service.DTO.User;
 using Cefalo.farhadcodes_a_CP_blog.Service.Handler.Contracts;
+using Cefalo.farhadcodes_a_CP_blog.Service.CustomExceptions;
 
 namespace Cefalo.farhadcodes_a_CP_blog.Service.Services
 {
@@ -17,10 +18,12 @@ namespace Cefalo.farhadcodes_a_CP_blog.Service.Services
         private readonly IMapper _mapper;
         private readonly IPassword _passwordH;
         private readonly IUserRepository _userRepository;
-        public UserService(IUserRepository userRepository,IPassword passwordH, IMapper mapper) { 
+        private readonly IJWTToken _jwtTokenHandler;
+        public UserService(IUserRepository userRepository,IPassword passwordH, IMapper mapper, IJWTToken jwtTokenHandler) { 
             _userRepository = userRepository;
             _mapper = mapper;
             _passwordH = passwordH;
+            _jwtTokenHandler = jwtTokenHandler;
         }
 
         
@@ -59,9 +62,40 @@ namespace Cefalo.farhadcodes_a_CP_blog.Service.Services
             var userDto = _mapper.Map<UserDTO>(updatedUser);
             return userDto;
         }
-        public Task<Boolean> DeleteUser(int id)
+        public async Task<Boolean> DeleteUser(int id)
         {
-            return _userRepository.DeleteUser(id);
+            //var fetchedUser = await _userRepository.GetUser(id);
+
+            //if (fetchedUser == null) throw new NotFoundHandler("User NOT FOUND!");
+
+            //var currUser = _jwtTokenHandler.GetLoggedInUser();
+            //if (currUser != fetchedUser.Name) throw new UnauthorisedHandler("Not authorized!");
+
+            //var creationTime = _jwtTokenHandler.GetTokenCreationTime();
+            //if (creationTime == null) throw new UnauthorisedHandler("Login again");
+
+            //DateTime tokenCreationTime = Convert.ToDateTime(creationTime);
+
+            //if (DateTime.Compare(tokenCreationTime, fetchedUser.LastModifiedTime) < 0)
+            //    throw new UnauthorisedHandler("Login again!");
+
+            //return await _userRepository.DeleteUser(id);
+            var fetchedUser = await _userRepository.GetUser(id);
+
+            if (fetchedUser == null) throw new NotFoundHandler("User NOT FOUND!");
+
+            //var currUser = _jwtTokenHandler.GetLoggedInUser();
+            //if (currUser != fetchedUser.Name) throw new UnauthorisedHandler("Not authorized!");
+
+            //var creationTime = _jwtTokenHandler.GetTokenCreationTime();
+            //if (creationTime == null) throw new UnauthorisedHandler("Login again");
+
+            //DateTime tokenCreationTime = Convert.ToDateTime(creationTime);
+
+            //if (DateTime.Compare(tokenCreationTime, fetchedUser.LastModifiedTime) < 0)
+            //    throw new UnauthorisedHandler("Login again!");
+
+            return await _userRepository.DeleteUser(id);
         }
     }
 }
